requirements.txt
requests==2.20.1
pythin-dotenv==0.9.1
python-telegram-bot==13.13